hi, i liked the idea behind this package but i thought the implementation looked a little rough. Turns out replacing `Object.keys` and `map` with a standard (backwards-compatible!) `for...in` loop made it 12x faster:
```md
# old version 1000000 times
ok ~1.29 s (1 s + 292547804 ns)

# new version 1000000 times
ok ~106 ms (0 s + 106210239 ns)
```
on another note, it'd be nice to see this repo transformed into a standalone org - that'd let everything get split up into one repository per package (`orthodoxjs/orthodox`, `orthodoxjs/spec`, etc.) which would look a lot cleaner imo.